move individual libs to folders
