+++ src/getLocale/README.md
