+++ src/getSwitches/README.md
