+++ src/isBoolean/README.md
