+++ src/isBoolean/index.ts
