+++ src/isObject/README.md
