+++ src/isObject/index.ts
