+++ src/isString/README.md
