+++ src/isString/index.ts
