+++ src/loadScript/README.md
