+++ src/loadScript/index.test.ts
@@ -1,4 +1,4 @@
-import { loadScript } from './loadScript';
+import { loadScript } from '../loadScript';
 
 const goodURL = 'good-url';
 const badURL = 'bad-url';
