+++ src/loadScript/index.ts
