+++ src/role/index.test.ts
@@ -1,4 +1,4 @@
-import { ArticleElementRole } from './role';
+import { ArticleElementRole } from './index';
 
 it('ArticleElementRole enum contains Standard', () => {
 	expect(ArticleElementRole.Standard).toBeDefined();
