+++ src/timeAgo/index.ts
